MarketSentimentAnalysisTool/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dash==1.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
DateTime==4.3
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
numpy==1.18.2
pandas==1.0.3
plotly==4.5.4
python-dateutil==2.8.1
pytz==2019.3
Quandl==3.5.0
requests==2.23.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
zope.interface==5.0.2
